added backend route checks

+++ backend/casts/api/views.py
@@ -9,22 +9,28 @@ from casts.api.serializers import CastCallSerializer
 #-----------------------------------------------------
 # local files import - Models
 from casts.models import Castcall
+from profiles.models import Profiles
+#-----------------------------------------------------
+# local files import - Permissions
+from casts.api.permissions import ReviewCastOrReadOnly
+from casts.api.permissionsCreate import CreateCastOrReadOnly
 
 
 
 
 class CastCallAV(APIView):
 
+    permission_classes = [CreateCastOrReadOnly]
+
     def get(self, request):
-        print("Request user: ", request.user)
-        print("Request ID: ", request.user.id)
+        
         castcall = Castcall.objects.all()
         serializer = CastCallSerializer(castcall, many=True)
         return Response(serializer.data)
 
     def post(self,request):
-        print("Request user: ", request.user)
-        print("Request ID: ", request.user.id)
+        profiles = Profiles.objects.all()
+        self.check_object_permissions(request, profiles)
         serializer = CastCallSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -37,9 +43,10 @@ class CastCallAV(APIView):
 
 class CastCallDetailAV(APIView):
 
+    permission_classes = [ReviewCastOrReadOnly]
+
     def get(self, request, pk):
-        print("Request user: ", request.user)
-        print("Request ID: ", request.user.id)
+        
         try:
             castcall = Castcall.objects.get(pk=pk)
         except Castcall.DoesNotExist:
@@ -48,9 +55,9 @@ class CastCallDetailAV(APIView):
         return Response(serializer.data)
 
     def put(self,request,pk):
-        print("Request user: ", request.user)
-        print("Request ID: ", request.user.id)
+        
         castcall = Castcall.objects.get(pk=pk)
+        self.check_object_permissions(request, castcall)
         serializer = CastCallSerializer(castcall, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -59,9 +66,8 @@ class CastCallDetailAV(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self,request,pk):
-        print("Request user: ", request.user)
-        print("Request ID: ", request.user.id)
         castcall = Castcall.objects.get(pk=pk)
+        self.check_object_permissions(request, castcall)
         castcall.delete()
         # return in form of status code
         return Response(status=status.HTTP_204_NO_CONTENT)
